% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_api_key.R
\name{create_api_key}
\alias{create_api_key}
\title{Setup API Credentials}
\usage{
create_api_key(dir = "private")
}
\arguments{
\item{dir}{Character. Directory to store credentials. Default is "private"}
}
\value{
Invisibly returns the path to the created YAML file
}
\description{
Creates a template YAML file for API credentials and adds it to .gitignore
}
