% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monitor.R
\name{get_monitor}
\alias{get_monitor}
\title{Retrieve NAPMD hourly monitor data for given station and variable}
\usage{
get_monitor(
  station_id = 453,
  variable = "pm25",
  start_time_utc = "2020-01-01 00:00:00",
  end_time_utc = "2020-12-31 00:00:00",
  username = api_key$username,
  password = api_key$password
)
}
\arguments{
\item{station_id}{Integer. NAPMD station ID number (see \code{\link{list_air_pollution_monitors}})}

\item{variable}{Character. Variable name (see \code{\link{get_variables}})}

\item{start_time_utc}{Character string in YYYY-MM-DD HH:MM:SS format (UTC timezone)}

\item{end_time_utc}{Character string in YYYY-MM-DD HH:MM:SS format (UTC timezone)}

\item{username}{NAPMD API username}

\item{password}{NAPMD API password}
}
\value{
data.table
}
\description{
Retrieve hourly monitor data of specified station and variable for range of dates from \code{start_time_utc} to \code{end_time_utc}, inclusive.
}
