% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monitor_daily.R
\name{get_monitor_daily}
\alias{get_monitor_daily}
\title{Retrieve NAPMD daily monitor data}
\usage{
get_monitor_daily(
  station_id = 300,
  variable = "pm25",
  start_date = "2020-01-01",
  end_date = "2020-02-01",
  username = api_key$username,
  password = api_key$password
)
}
\arguments{
\item{station_id}{Integer. NAPMD station ID number (see \code{\link{list_air_pollution_monitors}})}

\item{variable}{Character. Variable name (see \code{\link{get_variables_daily}})}

\item{start_date}{Date format or character string in YYYY-MM-DD format.}

\item{end_date}{Date format or character string in YYYY-MM-DD format.}

\item{username}{NAPMD API username}

\item{password}{NAPMD API password}
}
\value{
data.table
}
\description{
Retrieve daily monitor data from a date range (\code{start_date} to \code{end_date}, inclusive) for a specified station and variable.
}
