% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_slice_all.R
\name{get_slice_all}
\alias{get_slice_all}
\title{Retrieve slice of NAPMD hourly monitor data}
\usage{
get_slice_all(
  variable = "pm25",
  time_utc = "2019-11-03 15:00:00",
  state = "NSW",
  username = api_key$username,
  password = api_key$password
)
}
\arguments{
\item{variable}{Character. Variable name (see \code{\link{get_variables_daily}})}

\item{time_utc}{Character string in YYYY-MM-DD HH:MM:SS format (UTC timezone)}

\item{state}{Character. State abbreviation (one of "ACT", "NSW", "NT", "QLD", "SA", "TAS", "WA", "VIC").}

\item{username}{NAPMD API username}

\item{password}{NAPMD API password}
}
\value{
data.table
}
\description{
Retrieve daily monitor data from all stations for a specified state, variable and date.
}
