% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_slice_daily.R
\name{get_slice_daily}
\alias{get_slice_daily}
\title{Retrieve slice of NAPMD daily monitor data}
\usage{
get_slice_daily(
  variable = "pm25",
  date = "2012-01-01",
  state = "TAS",
  username = api_key$username,
  password = api_key$password
)
}
\arguments{
\item{variable}{Character. Variable name (see \code{\link{get_variables_daily}})}

\item{date}{Date format or character string in YYYY-MM-DD format.}

\item{state}{Character. State abbreviation (one of "ACT", "NSW", "NT", "QLD", "SA", "TAS", "WA", "VIC").}

\item{username}{NAPMD API username}

\item{password}{NAPMD API password}
}
\value{
data.table
}
\description{
Retrieve daily monitor data from all stations for a specified state, variable and date.
}
