% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_times.R
\name{get_times}
\alias{get_times}
\title{Get time range of monitor hourly data}
\usage{
get_times(
  station_id = 453,
  variable = "pm25",
  username = api_key$username,
  password = api_key$password
)
}
\arguments{
\item{station_id}{Integer. NAPMD station ID number (see \code{\link{list_air_pollution_monitors}})}

\item{variable}{Character. Variable name (see \code{\link{get_variables}})}

\item{username}{NAPMD API username}

\item{password}{NAPMD API password}
}
\value{
data.table
}
\description{
Retrieve time range of available hourly data for specified station and variable.
}
