% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_times_daily.R
\name{get_times_daily}
\alias{get_times_daily}
\title{Get time range of monitor daily data}
\usage{
get_times_daily(
  station_id = 300,
  variable = "pm25",
  username = api_key$username,
  password = api_key$password
)
}
\arguments{
\item{station_id}{Integer. NAPMD station ID number (see \code{\link{list_air_pollution_monitors}})}

\item{variable}{Character. Variable name (see \code{\link{get_variables_daily}})}

\item{username}{NAPMD API username}

\item{password}{NAPMD API password}
}
\value{
data.table
}
\description{
Retrieve time range of available daily data for specified station and variable.
}
