% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables.R
\name{get_variables}
\alias{get_variables}
\title{Get variables of monitor hourly data}
\usage{
get_variables(
  station_id = 453,
  username = api_key$username,
  password = api_key$password
)
}
\arguments{
\item{station_id}{Integer. NAPMD station ID number (see \code{\link{list_air_pollution_monitors}})}

\item{username}{NAPMD API username}

\item{password}{NAPMD API password}
}
\value{
data.table
}
\description{
Retrieve available variables of hourly data for specified station.
}
