% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables_daily.R
\name{get_variables_daily}
\alias{get_variables_daily}
\title{Get variables of monitor daily data}
\usage{
get_variables_daily(
  station_id = 300,
  username = api_key$username,
  password = api_key$password
)
}
\arguments{
\item{station_id}{Integer. NAPMD station ID number (see \code{\link{list_air_pollution_monitors}})}

\item{username}{NAPMD API username}

\item{password}{NAPMD API password}
}
\value{
data.table
}
\description{
Retrieve available variables and corresponding units of daily data for specified station.
}
