% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_air_pollution_monitors.R
\name{list_air_pollution_monitors}
\alias{list_air_pollution_monitors}
\title{Get monitor info}
\usage{
list_air_pollution_monitors(
  state = "ACT",
  username = api_key$username,
  password = api_key$password
)
}
\arguments{
\item{state}{Character. State abbreviation (one of "ACT", "NSW", "NT", "QLD", "SA", "TAS", "WA", "VIC").}

\item{username}{NAPMD API username}

\item{password}{NAPMD API password}
}
\value{
data.table
}
\description{
Retrieve station ID, name and spatial information for specified state. Latitudes and longitudes are given in GDA94 coordinate reference system (EPSG:4283).
}
