
\name{napmdtools-package}
\alias{napmdtools-package}
\alias{napmdtools}
\docType{package}
\title{
getPassword and connect2postgres
}
\description{
getPassword and connect2postgres~
}
\details{
\tabular{ll}{
Package: \tab napmdtools\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2023-11-03\cr
License: \tab GPL2\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Who wrote it

Maintainer: ivanhanigan <ivan.hanigan@gmail.com> and grantwilliamson

}
\references{

}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{

}
